DRY up Team and Track Stream Filters common code
Move common code into new class ActivityStreamFilters
