+++ lib/exercism.rb
@@ -6,6 +6,7 @@ require_relative '../config/flipper'
 
 require 'exercism/acl'
 require 'exercism/activity_stream'
+require 'exercism/activity_stream_filters'
 require 'exercism/assumable_user'
 require 'exercism/authentication'
 require 'exercism/client_version'
