+++ lib/exercism/track_stream_filters.rb
@@ -1,45 +1,7 @@
 class TrackStream
   class Filter
+    include ActivityStream::Filter
     attr_reader :viewer_id, :track_id
-
-    def items
-      @items ||= execute(items_sql).map do |row|
-        item(row["id"], row["total"])
-      end.sort(&order).each do |item|
-        item.unread = unread(item)
-      end
-    end
-
-    private
-
-    def unread(item)
-      [item.total - read(item.id), 0].max
-    end
-
-    def read(id)
-      views_by_id[id] + watermarked_by_id[id]
-    end
-
-    # By default, don't bother changing the order.
-    def order
-      proc { 0 }
-    end
-
-    def views_by_id
-      @views_by_id ||= execute(views_sql).each_with_object(Hash.new(0)) do |row, views|
-        views[row["id"]] = row["total"].to_i
-      end
-    end
-
-    def watermarked_by_id
-      @watermarked_by_id ||= execute(watermarks_sql).each_with_object(Hash.new(0)) do |row, watermarked|
-        watermarked[row["id"]] = row["total"].to_i
-      end
-    end
-
-    def execute(query)
-      ActiveRecord::Base.connection.execute(query).to_a
-    end
   end
 
   class TrackFilter < Filter
@@ -55,7 +17,7 @@ class TrackStream
     end
 
     # rubocop:disable Metrics/MethodLength
-    def items_sql
+    def sql
       <<-SQL
         SELECT ex.language AS id, COUNT(ex.id) AS total
         FROM user_exercises ex
@@ -142,7 +104,7 @@ class TrackStream
     end
 
     # rubocop:disable Metrics/MethodLength
-    def items_sql
+    def sql
       <<-SQL
         SELECT ex.slug AS id, COUNT(ex.id) AS total
         FROM user_exercises ex
@@ -227,7 +189,7 @@ class TrackStream
     end
 
     # rubocop:disable Metrics/MethodLength
-    def items_sql
+    def sql
       <<-SQL
         SELECT u.username AS id, COUNT(ex.id) AS total
         FROM user_exercises ex
